Relocate the flashes service to common folder so that login-app can use it as well
