+++ src/common/flashes/collection-view.js
