+++ src/common/flashes/collection.js
