+++ src/common/flashes/item-template.hbs
