+++ src/common/flashes/item-view.js
