+++ src/common/flashes/model.js
